Migrate convertFromThousandthInchToInch and formatToDimensionsInches over into utils/formatters
- use Number.isFinite instead of lodash isFinite
- return new object rather than doing parameter reassignment
