+++ src/shared/formatters.js
@@ -47,28 +47,6 @@ export function convertFromBaseQuantity(baseQuantity) {
   return baseQuantity / 10000;
 }
 
-// Format a thousandth of an inch into an inch, e.g. 16700 -> 16.7
-export function convertFromThousandthInchToInch(thousandthInch) {
-  if (!isFinite(thousandthInch)) {
-    return null;
-  }
-
-  return thousandthInch / 1000;
-}
-
-// Format a dimensions object length, width and height to inches
-export function formatToDimensionsInches(dimensions) {
-  if (!dimensions) {
-    return undefined;
-  }
-
-  dimensions.length = convertFromThousandthInchToInch(dimensions.length);
-  dimensions.width = convertFromThousandthInchToInch(dimensions.width);
-  dimensions.height = convertFromThousandthInchToInch(dimensions.height);
-
-  return dimensions;
-}
-
 // Format dimensions object length, width and height to base dimensions
 export function formatDimensionsToThousandthInches(dimensions) {
   if (!dimensions) {
